refactor: move common tests functions into a separate file
